pyproject.toml
[project]
dependencies = [
  "black>=24.0,<25.11",
  "django>=5.2.8",
  "python-dotenv>=1.0.0",
  "djangorestframework>=3.16.1",
  "psycopg2-binary>=2.9.9",
  "django-filter>=23.5",
  "ruff>=0.14.2",
  "gunicorn",
]
description = "Application for task management"
name = "alliance"
requires-python = ">=3.13"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "black>=24.0,<25.11",
  "ruff>=0.14.2",
]

[tool.ruff]
# Исключить стандартные директории
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-references",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]
indent-width = 4
# line-length = 88
target-version = "py313"

[tool.ruff.lint]
ignore = ["F401", "E501"] # Отключает проверку неиспользуемых импортов и длинных строк
select = ["E", "F"]

[tool.black]
# line-length = 88
target-version = ['py313']
